pyproject.toml
[project]
name = "pdfsensei"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.20",
    "langchain-huggingface>=1.2.1",
    "langchain-mistralai>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.0",
    "pdf2image>=1.17.0",
    "pdfminer-six>=20260107",
    "pi-heif>=1.3.0",
    "pillow>=12.1.1",
    "pypdf>=6.9.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "tqdm>=4.67.3",
    "unstructured[pdf]>=0.21.5",
    "unstructured-inference>=1.2.0",
    "pytesseract>=0.3.13",
    "langchain-google-genai>=4.2.1",
]
